refactor: refactor all folders and rename the Java classes
