+++ src/main/java/de/mikeyllp/miniGamesV4/commands/invites/InvitesRPSGameCommand.java
@@ -1,19 +1,19 @@
-package de.mikeyllp.miniGamesV4.commands.subCommands;
+package de.mikeyllp.miniGamesV4.commands.invites;
 
 import dev.jorel.commandapi.CommandAPICommand;
-import dev.jorel.commandapi.arguments.*;
-
-
+import dev.jorel.commandapi.arguments.ArgumentSuggestions;
+import dev.jorel.commandapi.arguments.StringArgument;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 import java.util.stream.Collectors;
 
-import static de.mikeyllp.miniGamesV4.map.InvitetPlayerHashMap.removeInvite;
-//adds all online players to the tab completer
-public class DeclinedMiniGamesCommand extends CommandAPICommand {
-    public DeclinedMiniGamesCommand(String commandName) {
+import static de.mikeyllp.miniGamesV4.storage.InvitePlayerStorage.canInvitePlayer;
+
+public class InvitesRPSGameCommand extends CommandAPICommand {
+    public InvitesRPSGameCommand(String commandName) {
         super(commandName);
+
         //This is a list that adds online players to the TabCompleter and you can´t use @
         withArguments(
                 new StringArgument("player")
@@ -23,21 +23,23 @@ public class DeclinedMiniGamesCommand extends CommandAPICommand {
                                         .collect(Collectors.toList())
                         ))
         );
-        executesPlayer((invited, args) -> {
-            Player inviter = Bukkit.getPlayerExact(args.get(0).toString());
+        //Send the Player a invite to play Rock Paper Scissors
+        executesPlayer((sender, args) -> {
+            Player targetPlayer = Bukkit.getPlayerExact(args.get(0).toString());
             String prefix = "<COLOR:DARK_GRAY>>> </COLOR><gradient:#00FF00:#007F00>MiniGames </gradient><COLOR:DARK_GRAY>| </COLOR>";
-            if (inviter == null) {
-                invited.sendRichMessage(prefix + "<red>Der Spieler ist nicht online.</red>");
+            if (targetPlayer == null) {
+                sender.sendRichMessage(prefix + "<red>Der Spieler ist nicht online.</red>");
                 return;
             }
             //Checks if the player has permission to use this command
-            if (!invited.hasPermission("minigamesv4.minigames")) {
-                invited.sendRichMessage(prefix + "<red>Du hast keine Berechtigung, um diesen Command zu nutzen.</red>");
+            if (!sender.hasPermission("minigamesv4.minigames")) {
+                sender.sendRichMessage(prefix + "<red>Du hast keine Berechtigung, um diesen Command zu nutzen.</red>");
                 return;
             }
 
-            //Remove this player from the invite list
-            removeInvite(inviter, invited);
+
+            canInvitePlayer(sender, targetPlayer, "RPS");
+
         });
     }
 }
